requirements.txt
Flask==2.3.2
Werkzeug==2.3.6
pymongo==4.4.1
python-dotenv==1.0.0
gunicorn==21.2.0
